+++ mandelbulber2/src/shader_background.cpp
@@ -78,7 +78,7 @@ sRGBAfloat cRenderWorker::BackgroundShader(const sShaderInputData &input) const
 				if (betaTexture > 0.5 * M_PI) betaTexture = 0.5 * M_PI - betaTexture;
 				if (betaTexture < -0.5 * M_PI) betaTexture = -0.5 * M_PI + betaTexture;
 
-				CVector2<double> tex(
+				CVector2<float> tex(
 					alphaTexture / (2.0 * M_PI) / params->backgroundHScale + params->backgroundTextureOffsetX,
 					(betaTexture / M_PI + 0.5) / params->backgroundVScale + params->backgroundTextureOffsetY);
 
@@ -107,7 +107,7 @@ sRGBAfloat cRenderWorker::BackgroundShader(const sShaderInputData &input) const
 				texX = (texX / params->backgroundHScale) + 0.5 + params->backgroundTextureOffsetX;
 				texY = (texY / params->backgroundVScale) + 0.5 + params->backgroundTextureOffsetY;
 
-				sRGBFloat pixel = data->textures.backgroundTexture.Pixel(CVector2<double>(texX, texY));
+				sRGBFloat pixel = data->textures.backgroundTexture.Pixel(CVector2<float>(texX, texY));
 				pixel2.R = pixel.R;
 				pixel2.G = pixel.G;
 				pixel2.B = pixel.B;
